requirements.txt
trimesh
tkinter
pillow